+++ sdk/identity/Azure.Identity/src/ManagedIdentityClientOptions.cs
@@ -21,5 +21,9 @@ namespace Azure.Identity
         public CredentialPipeline Pipeline { get; set; }
 
         public bool ExcludeTokenExchangeManagedIdentitySource { get; set; }
+
+        public bool EnableManagedIdentityLegacyBehavior { get; set; } = Environment.GetEnvironmentVariable("AZURE_IDENTITY_ENABLE_LEGACY_IMDS_BEHAVIOR") != null;
+
+        public bool IsForceRefreshEnabled { get; set; }
     }
 }